+++ src/Web/Http/HttpHandlerController.php
@@ -59,10 +59,9 @@ class HttpHandlerController
      */
     public function handle(Request $symfonyRequest, string $portalNodeId, string $path): Response
     {
-        $response = $this->httpHandleService->handle(
-            $this->getRequest($symfonyRequest, $path),
-            $this->getPortalNodeKey($portalNodeId)
-        );
+        $portalNodeKey = $this->getPortalNodeKey($portalNodeId);
+        $request = $this->getRequest($symfonyRequest, $path);
+        $response = $this->httpHandleService->handle($request, $portalNodeKey);
 
         return $this->httpFoundationFactory->createResponse($response);
     }
@@ -84,9 +83,20 @@ class HttpHandlerController
 
     protected function getRequest(Request $symfonyRequest, string $path): ServerRequestInterface
     {
-        $symfonyRequest->server = new ServerBag();
+        $serverBag = new ServerBag();
+        // needed for PSR HTTP Factory to set the query parameters in the URL
+        $serverBag->set('QUERY_STRING', $symfonyRequest->server->get('QUERY_STRING'));
+        $symfonyRequest->server = $serverBag;
         $request = $this->psrHttpFactory->createRequest($symfonyRequest);
 
+        foreach (\array_keys($request->getAttributes()) as $attributeKey) {
+            $request = $request->withoutAttribute($attributeKey);
+        }
+
+        $request = $this->withoutConnectionAndProxyHeaders($request);
+        $request = $this->withoutSymfonyHeaders($request);
+
+        $request = $request->withAttribute(HttpHandleServiceInterface::REQUEST_ATTRIBUTE_ORIGINAL_REQUEST, $request);
         $request = $request->withUri(
             $request->getUri()
             ->withScheme('')
@@ -97,10 +107,33 @@ class HttpHandlerController
 
         $request = $request->withoutHeader('host');
 
-        foreach (\array_keys($request->getAttributes()) as $attributeKey) {
-            $request = $request->withoutAttribute($attributeKey);
+        return $request;
+    }
+
+    private function withoutConnectionAndProxyHeaders(ServerRequestInterface $request): ServerRequestInterface
+    {
+        $headersToRemove = [
+            'connection',
+            'forwarded',
+            'proxy-connection',
+        ];
+
+        foreach (\array_keys($request->getHeaders()) as $headerName) {
+            // check every x-forwarded header to also support non "standard" headers from proxies like Traefik and AWS ELB
+            if (\preg_match('/^x[-_]forwarded[-_]/i', $headerName) === 1) {
+                $headersToRemove[] = $headerName;
+            }
+        }
+
+        foreach ($headersToRemove as $header) {
+            $request = $request->withoutHeader($header);
         }
 
         return $request;
+    }
+
+    private function withoutSymfonyHeaders(ServerRequestInterface $request): ServerRequestInterface
+    {
+        return $request->withoutHeader('x-php-ob-level');
     }
 }